requirements.txt
PyMuPDF
pdfplumber
langchain
langchain-community
sentence-transformers
faiss-cpu
Flask
pypdf




python-docx
pandas
openai-whisper
flask-socketio
pyjwt

